requirements.txt
ccxt
websockets
python-dotenv
pandas
pyyaml
